REFACTOR: devide imagequest.dart into controller and view
